+++ TradingApp.Core/Entitites/DailyTF.cs
@@ -2,7 +2,7 @@ namespace TradingApp.Core.Entities
 {
   public class DailyTF : Candle
   {
-    public DailyTF(int token, DateTime time, double open, double high, double low, double close, double volume)
+    public DailyTF(int token, DateTime time, double open, double high, double low, double close, int volume)
       : base(token, time, open, high, low, close, volume)
     {
     }
